pyproject.toml
[tool.poetry]
name = "squest"
version = "2.0.0b"
description = "Service catalog on top of Ansible Tower"
authors = ["Nicolas Marcq <nicolas.marcq@hpe.com>", "Elias Boulharts <elias.boulharts@hpe.com", "Anthony Belhadj <anthony.belhadj@hpe.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.13"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
mysqlclient = "^2.1.1"
PyYAML = "^6.0"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
towerlib = "^3.10.0"
Pillow = "^9.2.0"
django-fsm = "^2.8.0"
django-celery-beat = "^2.3.0"
Pygments = "^2.12.0"
django-filter = "^22.1"
coverage = "^6.4.1"
django-auth-ldap = "^4.1.0"
django-node-assets = "^0.9.11"
mkdocs = "^1.3.0"
mkdocs-material = "^8.3.9"
pymdown-extensions = "^9.5"
drf-yasg = "^1.20.0"
graphviz = "^0.20"
django-taggit = "^3.0.0"
mike = "^1.1.2"
martor = "^1.6.4"
django-tempus-dominus = "^5.1.2"
gunicorn = "^20.1.0"
django-tables2 = "^2.4.1"
django-dbbackup = "^3.3.0"
django-cleanup = "^6.0.0"
GitPython = "^3.1.27"
prometheus-client = "^0.14.1"
redis = "^4.3.4"
django-redis = "^5.2.0"
django-cachalot = "^2.5.1"
django-mysql = "^4.7.0"
psycopg2 = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
